+++ fv3core/fv3core/initialization/baroclinic.py
@@ -523,8 +523,8 @@ def init_baroclinic_state(
         sample_quantity.metadata.gt4py_backend,
     )
 
-    comm.halo_update(state.phis_quantity, n_points=nhalo)
+    comm.halo_update(state.phis, n_points=nhalo)
 
-    comm.vector_halo_update(state.u_quantity, state.v_quantity, n_points=nhalo)
+    comm.vector_halo_update(state.u, state.v, n_points=nhalo)
 
     return state